✨ [FE] feat: 감정 목록에 이미지 추가, 현재 지도의 위도/경도 범위 파악
